Move chain parity test to app module
The app test source set sees every chain module, so the test can
register all plugins as more chains are extracted.
